# sentinel-ai/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SentinelAI API
  version: 0.1.0
  description: |
    Implemented: the security-engine scan contract (`/v1/scan`, internal) and the gateway routes for security scan/check,
    AI chat/generate, file scanning (`/v1/files/scan`), events, usage and policies (API-key auth), login/JWT, user and
    team management (`/v1/users`, `/v1/invitations`, `/v1/teams`) and per-organization provider credentials
    (`/v1/providers`). Routes marked (planned) are specified ahead of implementation.

    Management routes (users, teams, invitations, providers, API keys) require a USER session (JWT): an API key is never
    accepted for them, so a leaked key cannot create users, change roles or store provider credentials.
    Wire format is snake_case JSON. Responses never contain matched sensitive values.
servers:
  - url: http://localhost:4000
    description: API gateway
  - url: http://localhost:8001
    description: Security engine (implemented, internal only)
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/scan:
    post:
      summary: "[internal, security-engine] Scan text and return decision, evidence, risk and sanitized text"
      security: [{ internalToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScanRequest" } } }
      responses:
        "200": { description: scan result, content: { application/json: { schema: { $ref: "#/components/schemas/ScanResult" } } } }
        "401": { description: invalid internal token }
        "422": { description: invalid request or policy (callers must treat as BLOCK) }
  /v1/replay:
    post:
      summary: |
        [internal, security-engine] Re-run a recorded decision with the ORIGINAL text and the event's stored explanation.
        Refuses text whose keyed hash differs from the recorded content_hmac; reuses the recorded judge verdict unless
        live_judge is true; read-only (no vault, no event). See services/security-engine/app/replay.py.
      security: [{ internalToken: [] }]
      responses: { "200": { description: "replay result (identical, differences, version diff)" }, "401": { description: invalid internal token } }
  /v1/security/scan:
    post:
      summary: Scan text under the organization policy (gateway; implemented)
      description: Requires permission `scan:use`. The scan is audited before the result is returned; if it cannot be audited the response is 503 and no text is returned.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GatewayScanRequest" } } } }
      responses:
        "200": { description: "decision + evidence + event_id (BLOCK also returns 200 with sanitized_text null)", content: { application/json: { schema: { $ref: "#/components/schemas/ScanResult" } } } }
        "401": { description: unauthorized }
        "403": { description: forbidden }
        "413": { description: payload too large }
        "422": { description: invalid request }
        "503": { description: audit unavailable (fail closed) }
...
  /v1/files/scan:
    post:
      summary: Scan an uploaded file (implemented)
      description: |
        Body is the RAW file bytes (`Content-Type: application/octet-stream`), not multipart. Only the extension of `X-Filename` is used;
        the name itself is never forwarded, stored or logged. Pipeline: type validation (magic bytes) -> malware scan -> extraction/OCR ->
        the same PII/secret/injection engine and org policy as text -> decision. BLOCK is a decision (HTTP 200), not an error, and a blocked
        result never carries text. The upload is held in memory only. Any failure to reach a safe verdict is a fail-closed BLOCK.
      security: [{ apiKeyAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: X-Filename, in: header, required: false, schema: { type: string }, description: "URL-encoded; only the extension is used" }
        - { name: X-Application, in: header, required: false, schema: { type: string } }
        - { name: X-Team, in: header, required: false, schema: { type: string } }
        - { name: X-Environment, in: header, required: false, schema: { type: string } }
      requestBody:
        required: true
        content: { application/octet-stream: { schema: { type: string, format: binary } } }
      responses:
        "200":
          description: Verdict
          content:
            application/json:
              schema:
                type: object
                required: [event_id, decision, blocked, failed_closed, reason, file, findings, risk, detections, sanitized_text, policy_id]
                properties:
                  event_id: { type: [string, "null"] }
                  decision: { $ref: "#/components/schemas/Action" }
                  blocked: { type: boolean }
                  failed_closed: { type: boolean }
                  reason: { type: [string, "null"], description: "e.g. macros_present, pdf_active_content, malware_detected, scanner_unreachable" }
                  file:
                    type: object
                    properties:
                      sha256: { type: string }
                      size: { type: integer }
                      detected_type: { type: [string, "null"] }
                      mime: { type: [string, "null"] }
                      pages: { type: [integer, "null"] }
                      ocr_used: { type: boolean }
                  findings:
                    type: array
                    items: { type: object, properties: { type: { type: string }, severity: { type: string }, detail: { type: string } } }
                  risk: { type: object }
                  detections: { type: array, items: { type: object } }
                  sanitized_text: { type: [string, "null"], description: "Null whenever blocked" }
                  policy_id: { type: string }
        "401": { description: unauthorized }
        "403": { description: forbidden }
        "413": { description: file larger than MAX_FILE_BYTES }
        "415": { description: unsupported content type }
        "422": { description: empty body or invalid headers }
        "429": { description: rate limited (30 uploads per minute per client) }
        "503": { description: the scan could not be audited, so no verdict is released }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKeyAuth: { type: apiKey, in: header, name: X-Sentinel-Api-Key }
    internalToken: { type: apiKey, in: header, name: X-Internal-Token }
  schemas:
    Severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    Action: { type: string, enum: [ALLOW, HASH, MASK, TOKENIZE, REDACT, QUARANTINE, BLOCK] }
    Direction: { type: string, enum: [INPUT, OUTPUT] }
    RiskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    EntityType:
      type: string
      enum: [NAME, LOCATION, EMAIL, PHONE, ADDRESS, DATE_OF_BIRTH, PAN, AADHAAR, PASSPORT, SSN, DRIVER_LICENSE, CREDIT_CARD, BANK_ACCOUNT, UPI, IFSC,
             API_KEY, AWS_CREDENTIAL, GOOGLE_CREDENTIAL, GITHUB_TOKEN, JWT, OAUTH_TOKEN, PASSWORD, PRIVATE_KEY, CONNECTION_STRING,
             HIGH_ENTROPY_SECRET, INTERNAL_URL, CONFIDENTIAL_MARKER, CUSTOM_CONFIDENTIAL,
             PROMPT_INJECTION, SYSTEM_PROMPT_EXTRACTION, JAILBREAK, DATA_EXFILTRATION]
    RequestContext:
      type: object
      additionalProperties: false
      properties:
        user_id: { type: string, maxLength: 256 }
        team: { type: string, maxLength: 256 }
        application: { type: string, maxLength: 256 }
        provider: { type: string, maxLength: 64 }
        model: { type: string, maxLength: 128 }
        environment: { type: string, maxLength: 64 }
        ip: { type: string, maxLength: 64 }
    TimeWindow:
      type: object
      additionalProperties: false
      required: [start, end]
      properties: { start: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }, end: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" } }
    RuleScope:
      type: object
      additionalProperties: false
      properties:
        users: { type: array, items: { type: string } }
        teams: { type: array, items: { type: string } }
        applications: { type: array, items: { type: string } }
        providers: { type: array, items: { type: string } }
        models: { type: array, items: { type: string } }
        environments: { type: array, items: { type: string } }
        ip_cidrs: { type: array, items: { type: string } }
        time_window_utc: { $ref: "#/components/schemas/TimeWindow" }
        directions: { type: array, items: { $ref: "#/components/schemas/Direction" } }
    PolicyRule:
      type: object
      additionalProperties: false
      required: [entity, action]
      description: ALLOW is rejected for credentials, payment cards and threat entities.
      properties:
        entity: { $ref: "#/components/schemas/EntityType" }
        action: { $ref: "#/components/schemas/Action" }
        severity: { $ref: "#/components/schemas/Severity" }
        min_confidence: { type: number, minimum: 0, maximum: 1, default: 0 }
        scope: { $ref: "#/components/schemas/RuleScope" }
    Policy:
      type: object
      additionalProperties: false
      required: [policy_id]
      properties:
        policy_id: { type: string, minLength: 1, maxLength: 128 }
        organization_id: { type: string, maxLength: 128 }
        version: { type: integer, minimum: 1, default: 1 }
        rules: { type: array, maxItems: 500, items: { $ref: "#/components/schemas/PolicyRule" } }
    ScanRequest:
      type: object
      additionalProperties: false
      required: [text, organization_id]
      properties:
        text: { type: string }
        direction: { $ref: "#/components/schemas/Direction" }
        organization_id: { type: string, minLength: 1, maxLength: 128 }
        context: { $ref: "#/components/schemas/RequestContext" }
        policy: { $ref: "#/components/schemas/Policy" }
    Detection:
      type: object
      description: Evidence only. Never contains the matched value.
      required: [entity, confidence, severity, location, detector, detector_version]
      properties:
        entity: { $ref: "#/components/schemas/EntityType" }
        confidence: { type: number, minimum: 0, maximum: 1 }
        severity: { $ref: "#/components/schemas/Severity" }
        location: { type: object, required: [start, end], properties: { start: { type: integer }, end: { type: integer } } }
        detector: { type: string }
        detector_version: { type: string }
        value_digest: { type: [string, "null"], description: keyed HMAC prefix for correlation }
    RiskFactor:
      type: object
      required: [name, contribution, detail]
      properties: { name: { type: string }, contribution: { type: number }, detail: { type: string } }
    Risk:
      type: object
      required: [risk_score, risk_level, decision, factors]
      properties:
        risk_score: { type: integer, minimum: 0, maximum: 100 }
        risk_level: { $ref: "#/components/schemas/RiskLevel" }
        decision: { $ref: "#/components/schemas/Action" }
        factors: { type: array, items: { $ref: "#/components/schemas/RiskFactor" } }
    EntityAction:
      type: object
      required: [entity, action, count]
      properties: { entity: { $ref: "#/components/schemas/EntityType" }, action: { $ref: "#/components/schemas/Action" }, count: { type: integer } }
    ScanResult:
      type: object
      required: [request_id, decision, failed_closed, detections, entity_actions, risk, sanitized_text, policy_id, detector_version, latency_ms]
      properties:
        request_id: { type: string }
        decision: { $ref: "#/components/schemas/Action" }
        failed_closed: { type: boolean }
        fail_closed_reason: { type: [string, "null"] }
        detections: { type: array, items: { $ref: "#/components/schemas/Detection" } }
        entity_actions: { type: array, items: { $ref: "#/components/schemas/EntityAction" } }
        risk: { $ref: "#/components/schemas/Risk" }
        sanitized_text: { type: [string, "null"], description: null whenever the decision withholds content (BLOCK/QUARANTINE) }
        policy_id: { type: string }
        detector_version: { type: string }
        latency_ms: { type: number }
        explanation: { oneOf: [{ $ref: "#/components/schemas/Explanation" }, { type: "null" }] }
    Explanation:
      type: object
      description: |
        Why a decision was made. No content: detector names, scores, versions and a keyed HMAC of the scanned text. The
        judge's free-text `reason` is returned by scans but never stored (gateway strips it; a DB CHECK rejects it).
      required: [decided_by, tier, detectors_fired, policy, versions, content_hmac]
      properties:
        decided_by: { enum: [rules, classifier, judge, fail_closed] }
        tier: { type: integer, minimum: 1, maximum: 3 }
        detectors_fired: { type: array, items: { type: object, properties: { detector: { type: string }, entity: { type: string }, count: { type: integer }, max_confidence: { type: number }, tier: { type: integer } } } }
        classifier: { type: [object, "null"], properties: { model: { type: string }, score: { type: number }, threshold: { type: number }, band_low: { type: [number, "null"] }, band_high: { type: [number, "null"] }, band: { enum: [attack, uncertain, benign] }, windows_scored: { type: integer }, windows_total: { type: integer } } }
        judge: { type: [object, "null"], properties: { called: { type: boolean }, cached: { type: boolean }, skipped_reason: { type: [string, "null"] }, verdict: { enum: [attack, benign, null] }, category: { type: [string, "null"] }, confidence: { type: [number, "null"] }, reason: { type: [string, "null"] }, model: { type: [string, "null"] }, prompt_version: { type: [string, "null"] }, latency_ms: { type: [number, "null"] } } }
        policy: { type: object, properties: { policy_id: { type: string }, policy_version: { type: integer }, deciding_entity: { type: [string, "null"] }, deciding_action: { type: string }, source: { enum: [policy_rule, baseline, risk_escalation, no_detection, fail_closed] } } }
        versions: { type: object, additionalProperties: { type: string } }
        content_hmac: { type: string, pattern: "^[0-9a-f]{64}$" }
...
    GatewayScanRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string }
        direction: { $ref: "#/components/schemas/Direction" }
        context: { type: object, additionalProperties: false, properties: { application: { type: string }, team: { type: string }, environment: { type: string }, model: { type: string } } }
...